refactor: create CompactBlockUtils and move functions
Create a CompatBlockUtils which lets us move some logic from PeerNetwork
class without separating the functions since they are similar in use.
This lets us remove some bloat from the quite large PeerNetwork class.
