+++ ironfish/src/utils/index.ts
@@ -8,6 +8,7 @@ export * from './bench'
 export * from './bigint'
 export * from './blockchain'
 export * from './buffer'
+export * from './compactblock'
 export * from './currency'
 export * from './enums'
 export * from './error'
@@ -15,8 +16,8 @@ export * from './file'
 export * from './graffiti'
 export * from './hash'
 export * from './json'
-export * from './math'
 export * from './map'
+export * from './math'
 export * from './node'
 export * from './promise'
 export * from './string'
